feat: restructure connections routing and views, replacing perfil with conexiones
