--- multi-stock-sync/src/views/Sync/Views/Perfil/Router/RouterPerfil.tsx
@@ -1,13 +1,0 @@
-import { Navigate, Route, Routes } from "react-router-dom";
-import HomePerfil from "../Home/HomePerfil";
-
-function RouterPerfil(){
-    return(
-        <Routes>
-            <Route path="/home" element={<HomePerfil/>}></Route>
-
-            <Route path="/*" element={<Navigate to="/sync/perfil/home"/>}/>
-        </Routes>
-    );
-};
-export default RouterPerfil;
